requirements.txt
python-dotenv
jupyterlab
ipywidgets
requests
numpy
pandas
scipy
scikit-learn
matplotlib
gensim
torch
transformers
tqdm
openai
gradio
langchain
tiktoken
faiss-cpu
langchain-openai
langchain_experimental
langchain_chroma
langchain[docarray]
datasets
sentencepiece
matplotlib
google-generativeai
anthropic
scikit-learn
unstructured
chromadb
plotly
jupyter-dash
beautifulsoup4
pydub
modal
ollama
accelerate
sentencepiece
bitsandbytes
psutil
setuptools
speedtest-cli
sentence_transformers
feedparser
google-genai
mcp[cli]
ipykernel
httpx
fastmcp